% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip images vertically}
\usage{
flip(images, frames = NULL)
}
\arguments{
\item{images}{an object of class \code{magick-image} to modify}

\item{frames}{integer vector of frame indices to duplicate. Defaults to
\code{NULL}, which duplicates all frames.}
}
\value{
a \code{magick-image} object
}
\description{
Flip images vertically
}
\section{Verbosity}{

  After each operation a message listing the updated frame sequence is
  printed in interactive sessions.  Use
  \code{stopmotion_verbosity(FALSE)} to suppress these messages, or set
  \code{options(stopmotion.verbose = FALSE)} in your script or
  \file{.Rprofile}.
}

\examples{
\dontrun{
  dino_dir <- system.file("extdata", package = "stopmotion")
  images <- read(dir = dino_dir)
  flip(images = images, frames = 2:3)
}
}
