% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read images into a stop-motion film}
\usage{
read(dir, pattern = "")
}
\arguments{
\item{dir}{path to directory containing the images relative to working
directory.}

\item{pattern}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}
}
\value{
an object of class \code{magick-image}
}
\description{
Reads all image files from \code{dir} (optionally filtered by
\code{pattern}) and returns them as a \code{magick-image} object.
}
\section{Frame order}{

  Frames are loaded in the order returned by \code{\link[base]{list.files}},
  which sorts filenames lexicographically.  This means the filesystem
  filename order determines the stop-motion frame order.  Name your files
  accordingly (e.g. \code{frame_001.png}, \code{frame_002.png}, \ldots) to
  guarantee the intended sequence.  If you need to reorder frames after
  loading, use \code{\link{arrange}}.
}

\examples{
 \dontrun{
   dino_dir <- system.file("extdata", package = "stopmotion")
   images <- read(dir = dino_dir)
}
}
