% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate images}
\usage{
rotate(images, degrees, frames = NULL)
}
\arguments{
\item{images}{an object of class \code{magick-image} to modify}

\item{degrees}{a number in \code{[-360, 360]} specifying the rotation angle.}

\item{frames}{integer vector of frame indices to duplicate. Defaults to
\code{NULL}, which duplicates all frames.}
}
\value{
a \code{magick-image} object
}
\description{
Replaces selected frames with a rotated version in place. For large-angle
transformations. For a small-angle hand-held rock effect, see
\code{\link{wiggle}}.
}
\section{Verbosity}{

  After each operation a message listing the updated frame sequence is
  printed in interactive sessions.  Use
  \code{stopmotion_verbosity(FALSE)} to suppress these messages, or set
  \code{options(stopmotion.verbose = FALSE)} in your script or
  \file{.Rprofile}.
}

\examples{
\donttest{
  dino_dir <- system.file("extdata", package = "stopmotion")
  images <- read(dir = dino_dir)
  rotate(images = images, degrees = 90, frames = 2L)
}
}
