% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{splice}
\alias{splice}
\title{Splice frames into a film}
\usage{
splice(images, insert, after)
}
\arguments{
\item{images}{an object of class \code{magick-image} to modify}

\item{insert}{an object of class \code{magick-image} containing 1 or more
images to be inserted.}

\item{after}{integer scalar (or vector of scalars) giving the frame
number(s) after which \code{insert} will be inserted.  When a vector is
supplied the insertions are applied left-to-right, each offset by the
cumulative growth of the film from prior insertions.}
}
\value{
a \code{magick-image} object
}
\description{
Splice frames into a film
}
\section{Verbosity}{

  After each operation a message listing the updated frame sequence is
  printed in interactive sessions.  Use
  \code{stopmotion_verbosity(FALSE)} to suppress these messages, or set
  \code{options(stopmotion.verbose = FALSE)} in your script or
  \file{.Rprofile}.
}

\examples{
\dontrun{
  dino_dir <- system.file("extdata", package = "stopmotion")
  images <- read(dir = dino_dir)
  splice(images = images, insert = magick::wizard, after = 1)
}
}
